问题:  flutter 类型转换修改  getRemoteHistoryMessages 返回值修改 原因及方案: Reviewer: 来源:

+++ android/src/main/java/com/example/rongcloud_im_plugin/RCIMFlutterWrapper.java
@@ -913,7 +913,7 @@ public class RCIMFlutterWrapper {
             String targetId = (String)map.get("targetId");
             Integer st = (Integer) map.get("sendStatus");
             Message.SentStatus sendStatus = Message.SentStatus.setValue(st.intValue());
-            Long sendTime = (Long)map.get("sendTime");
+            Integer sendTime = (Integer)map.get("sendTime");
 
             String contentStr = (String)map.get("content");
 
@@ -976,7 +976,7 @@ public class RCIMFlutterWrapper {
             Integer st = (Integer)map.get("rececivedStatus");
             Message.ReceivedStatus receivedStatus = new Message.ReceivedStatus(st.intValue());
             String senderUserId = (String)map.get("senderUserId");
-            Long sendTime = (Long)map.get("sendTime");
+            Integer sendTime = (Integer)map.get("sendTime");
 
             String contentStr = (String)map.get("content");
 
@@ -1035,7 +1035,7 @@ public class RCIMFlutterWrapper {
             Integer t = (Integer)map.get("conversationType");
             Conversation.ConversationType type = Conversation.ConversationType.setValue(t.intValue());
             final String targetId = (String)map.get("targetId");
-            Long recordTime = (Long)map.get("recordTime");
+            Integer recordTime = (Integer)map.get("recordTime");
             Integer count = (Integer)map.get("count");
 
             RongIMClient.getInstance().getRemoteHistoryMessages(type, targetId, recordTime.longValue(), count, new RongIMClient.ResultCallback<List<Message>>() {
